fix: Look up ContactId::DEVICE by ContactId::DEVICE_ADDR (#6323)
At least this fixes the bug when the device chat is e.g. archived by the user on one device, but
this user action is synchronized as archiving a chat with a new contact having the address
"device@localhost" on another device. Not sure if we want to sinchronize actions on the device chats
at all because they are chats with different devices, but maybe that simplifies UX, so let's discuss
it first and maybe disable synchronization in a separate commit.