fix: Ignore protected headers in outer message part (#6357)
Delta Chat always adds protected headers to the inner encrypted or signed message, so if a protected
header is only present in the outer part, it should be ignored because it's probably added by the
server or somebody else. The exception is Subject because there are known cases when it's only
present in the outer message part, e.g. an encrypted unsigned Thunderbird message.

Also handle "Auto-Submitted" and "Autocrypt-Setup-Message" as protected headers on the receiver
side, this was apparently forgotten. This may fix #6357 where Saved Messages
(i.e. `ContactId::SELF`) and some contacts are unexpectedly marked as bots which can happen if
e.g. the server adds "Auto-Submitted: auto-generated" to messages for some reason. Maybe sounds
unlikely, but let's try.